+++ J2018112664江胤佐1/src/main/java/pers/jiangyinzuo/rollcall/dao/TeachingClassDao.java
@@ -12,7 +12,7 @@ public interface TeachingClassDao {
 	void insertTeachingClass(TeachingClass teachingClass) throws IOException, IllegalAccessException, IllegalArgumentException,
 	InvocationTargetException, NoSuchMethodException, SecurityException;;
 	
-	List<TeachingClass> queryTeachingClassesByStudentId(Integer studentId) throws FileNotFoundException, ClassNotFoundException, IOException, CustomException;
+	List<TeachingClass> queryTeachingClassesByStudentId(Long studentId) throws FileNotFoundException, ClassNotFoundException, IOException, CustomException;
 	
-	List<TeachingClass> queryTeachingClassesByTeacherId(Integer teacher) throws FileNotFoundException, ClassNotFoundException, IOException, CustomException;
+	List<TeachingClass> queryTeachingClassesByTeacherId(Long teacher) throws FileNotFoundException, ClassNotFoundException, IOException, CustomException;
 }
