+++ J2018112664江胤佐1/src/main/java/pers/jiangyinzuo/rollcall/service/LoginService.java
@@ -8,7 +8,7 @@ import main.java.pers.jiangyinzuo.rollcall.entity.Student;
 import main.java.pers.jiangyinzuo.rollcall.entity.Teacher;
 
 public interface LoginService {
-	Teacher teacherLogin(Integer teacherId, String pwd) throws CustomException, FileNotFoundException, IOException, ClassNotFoundException;
+	Teacher teacherLogin(Long teacherId, String pwd) throws CustomException, FileNotFoundException, IOException, ClassNotFoundException;
 
-	Student studentLogin(Integer studentId, String pwd) throws CustomException, FileNotFoundException, IOException, ClassNotFoundException;
+	Student studentLogin(Long studentId, String pwd) throws CustomException, FileNotFoundException, IOException, ClassNotFoundException;
 }
