+++ J2018112664江胤佐1/src/main/java/pers/jiangyinzuo/rollcall/service/TeachingClassService.java
@@ -5,13 +5,12 @@ import java.io.IOException;
 import java.util.List;
 
 import main.java.pers.jiangyinzuo.rollcall.common.CustomException;
-import main.java.pers.jiangyinzuo.rollcall.entity.Student;
 import main.java.pers.jiangyinzuo.rollcall.entity.TeachingClass;
 
 public interface TeachingClassService {
-	List<TeachingClass> queryTeachingClassesByStudentId(Integer studentId)
+	List<TeachingClass> queryTeachingClassesByStudentId(Long studentId)
 			throws FileNotFoundException, ClassNotFoundException, IOException, CustomException;
 
-	List<TeachingClass> queryTeachingClassesByTeacherId(Integer teacherId)
+	List<TeachingClass> queryTeachingClassesByTeacherId(Long teacherId)
 			throws FileNotFoundException, ClassNotFoundException, IOException, CustomException;
 }
